Church/requirements-dev.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
dj-database-url==2.3.0
Django==5.1.3
django-bootstrap5==24.3
django-environ==0.11.2
django-mdeditor==0.1.20
django-ratelimit==4.1.0
filelock==3.16.1
fonttools==4.55.0
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
huggingface-hub==0.26.3
idna==3.10
jiter==0.8.0
kiwisolver==1.4.7
Markdown==3.7
matplotlib==3.9.2
numpy==2.1.3
openai==1.55.3
packaging==24.2
pillow==11.0.0
pydantic==2.10.2
pydantic_core==2.27.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.2
stripe==11.3.0
tokenizers==0.20.3
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.12.2
urllib3==2.2.3
wordcloud==1.9.4
